add avatar in view
